优惠券接口优化

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/coupon/CouponParam.java
@@ -124,7 +124,7 @@ public class CouponParam {
     /**
      * 优惠券有效期类型
      */
-    private Integer    validityType;
+    private Byte    validityType;
     /**
      * 优惠券有效天数
      */
